pyproject.toml
[project]
name = "ex16-knn-algopractice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "openai>=2.7.1",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "tqdm>=4.67.1",
    "umap-learn>=0.5.9.post2",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "pytest>=8.4.2",
    "ruff>=0.14.3",
]
